requirements.txt
Django==2.2.3
pytz==2019.1
sqlparse==0.3.0
django-taggit==1.1.0
django-cors-headers==3.0.2